Products GETs opened

+++ src/products/products.controller.ts
@@ -6,12 +6,15 @@ import {
   Param,
   Patch,
   Post,
+  Req,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Types } from 'mongoose';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductsService } from './products.service';
 import { ApiBearerAuth } from '@nestjs/swagger';
+import { Request } from 'express';
 
 @ApiBearerAuth()
 @Controller('products')
@@ -29,7 +32,10 @@ export class ProductsController {
   }
 
   @Post()
-  async create(@Body() body: CreateProductDto) {
+  async create(@Body() body: CreateProductDto, @Req() req: Request) {
+    if (!req['isAdmin']) {
+      throw new UnauthorizedException();
+    }
     return await this.productsService.create(body);
   }
 
@@ -37,12 +43,20 @@ export class ProductsController {
   async update(
     @Param('id') id: Types.ObjectId | string,
     @Body() body: UpdateProductDto,
+    @Req() req: Request,
   ) {
+    if (!req['isAdmin']) {
+      throw new UnauthorizedException();
+    }
     return await this.productsService.update(id, body);
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: Types.ObjectId | string) {
+  async remove(@Param('id') id: Types.ObjectId | string, @Req() req: Request) {
+    if (!req['isAdmin']) {
+      throw new UnauthorizedException();
+    }
+
     return await this.productsService.remove(id);
   }
 }
